requirements.txt
flake8==3.6.0
Flask==1.0.2
pygcode==0.2.1
pyserial==3.4
RPi.GPIO==0.6.3
svgwrite==1.1.12
typeguard==2.7.1
unittest2==1.1.0
